added progress bar and detail screen
